Se agrega autenticacion y validacion de usuario con token

+++ src/controllers/projectController.ts
@@ -5,10 +5,10 @@ export class ProjectController {
   //--------POST--------//
   static createProject = async (req: Request, res: Response) => {
     const project = new Project(req.body);
-
+    //Asigna un manager
+    project.manager = req.user.id;
     try {
       await project.save();
-
       res.send("¡Proyecto creado correctamente!");
     } catch (error) {
       res.status(500).json({ error: "Hubo un error!" });
@@ -18,8 +18,9 @@ export class ProjectController {
   //--------GET--------//
   static getAllProjects = async (req: Request, res: Response) => {
     try {
-      const projects = await Project.find({});
-
+      const projects = await Project.find({
+        $or: [{ manager: { $in: req.user.id } }],
+      });
       res.json(projects);
     } catch (error) {
       res.status(500).json({ error: "Hubo un error" });
@@ -35,7 +36,10 @@ export class ProjectController {
         const error = new Error("¡Proyecto no encontrado!");
         return res.status(404).json({ error: error.message });
       }
-
+      if (project.manager.toString() !== req.user.id.toString()) {
+        const error = new Error("¡Accion no valida!");
+        return res.status(404).json({ error: error.message });
+      }
       res.json(project);
     } catch (error) {
       console.log(error);
@@ -50,6 +54,12 @@ export class ProjectController {
 
       if (!project) {
         const error = new Error("¡Proyecto no encontrado!");
+        return res.status(404).json({ error: error.message });
+      }
+      if (project.manager.toString() !== req.user.id.toString()) {
+        const error = new Error(
+          "¡Solo el Manager puede actualizar un proyecto!"
+        );
         return res.status(404).json({ error: error.message });
       }
       project.projectName = req.body.projectName;
@@ -72,7 +82,10 @@ export class ProjectController {
         const error = new Error("¡Proyecto no encontrado!");
         return res.status(404).json({ error: error.message });
       }
-
+      if (project.manager.toString() !== req.user.id.toString()) {
+        const error = new Error("¡Solo el Manager puede eliminar un proyecto!");
+        return res.status(404).json({ error: error.message });
+      }
       await project.deleteOne();
       res.send("¡Proyecto eliminado correctamente!");
     } catch (error) {
